Completely reviewed constructors and their bridges.
Now for each class we have *two* constructors, which share
the same 'prototype' property. One is for use by Scala.js,
the other is the JS bridge.

Both are used with the 'new' keyword, which is optimized
better than Object.create in most implementations.

Also, both can be used for instanceof tests in JavaScript,
if necessary.

Currently prototype.constructor is set to the Scala.js
constructor. This should probably be reviewed.