package-js now brings in .js files of dependencies.
This means that issuing package-js should only be done
for the final, standalone application. It will fetch
all individual .js files from all the dependencies, as
well as the application; sort them appropriately, and
package them in one single .js file.

The -dev.html of application should therefore not
include scalajs-runtime.js anymore, nor should it
declare it as an input for optimize-js. For the same
reason, manual copying of scalajs-runtime.js is not
needed anymore (and should not be done).