Fix encoding of method syms owned by impl classes.
Previously, I thought that signatures of methods
in impl classes were globally unique, because their
first parameter is of the type of the corresponding
interface.

But if the trait has a self type, the type of this
parameter is the self type. Given that two traits
can have the same self type, they can have methods
with the same signature.

So now the encoding of a method sym owned by an
impl class is prefixed by the full name of its
owning impl class.