Fix #39: js.Array of Int causes ClassCastExceptions.
Actually the problem was more general: giving values of
Scala primitive types to a JS method with varargs of a
generic type (or Any) caused the values not to be boxed.

This commit includes a non-optimal fix that basically
discards the "optimization" of expanding the actual
arguments when they are primitive value types. Instead
it uses js.Any.fromTraversableOnce and calls apply on
the method.

A proper implementation should test when the values must
be boxed in that case. But figuring this out with the
information we have after erasure is tricky.