Always add snapshots resolver in Scala.js settings.
It is going to be some time before Scala.js libraries
are stable enough to have releases (like DOM and jQuery
facade types), so we'd better not annoy users with
having to add an additional resolver when they are
using a released version of Scala.js itself.

This should probably be reverted once we reach a 1.0,
but that is not for tomorrow.