Make sure js.Block() is always created with 2+ elements.
If applied to 0 element, it becomes js.Skip(). If applied
to 1 element, that element is returned.

This makes sure that trees reaching JSDesugaring are as
unnested as possible. This, in turn, allows it to consider
as many trees as possible as being already simple JavaScript
expressions, which need not be unnested with temporary
variables.