Alternative implem for protection of receivers of raw apply.
This is an alternative implementation of the protection
mechanism introduced in 3be49088aaee3ec8d4cdabf3ef286e342febcd2e.

This new mechanism goes through a simple identity function,
instead of going through a local variable. It produces less
code because 1) it is shorter in the first place and 2) it
does not introduce a non-JS-expression where there was none,
and hence the desugaring does not have to extract even more
temporary variables.