Fix #79: Allow anonymous subclasses of js.Any for lambdas
Allows anonymous classes that extend js.FunctionN in order to allow
for lambda expressions. Further, allows anonymous classes extending
js.FunctionN to be nested within classes extending js.Any. This could
theoretically be abused to access the this reference of the outer
instance (which doesn't work).

This is a fix required for the SAM treatment to work.