Fix #303: Reflective calls to methods on java.lang.Object
Reflective calls to java.lang.Object made the Scala.js compiler crash
because of the special handling for strings in reflective calls. This
commit fixes the issue. It also adds reflective call proxies to the JS
implementation of java.lang.Object.

Further, it emits an error message if a reflective call is attempted
on a method of java.lang.Object that is not supported by Scala.js
(e.g. wait).