Simple test of testing example under (pre)optimizeJS.
It works for the testing example because it is small. We
cannot do the same thing for the test suite because Rhino
crashes under the load. We'll have to fix this in a later
version of the sbt plugin, but at least this commit makes
sure we won't break inherent support for testing of
preoptimizeJS and optimizeJS.