Fix #338: optionally retarget @JSExport.
The "API" to do so is to define a variable named
__ScalaJSExportsNamespace in the scope (typically the global
scope) before code emitted by Scala.js is executed. If defined
and an object, exports of classes and objects will be stored
in that object. If it is not defined, they are stored in the
global scope, as they used to be.