Make packageJS work directly on jars.
This is a second step to #392. Now packageJS also reads its
inputs directly from jars on the classpath, rather than requiring
everything to be extracted on disk.

As was the case for preoptimizeJS, this makes no-op packageJS
being virtually immediate on Windows. Since packageJS is decomposed
in three parts, this is particularly important as the extdeps part
is almost always a no-op, and the two other parts are fast because
they do not operate on much data to begin with.