Do not box Floats.
Boxed Floats and Doubles are now complete indistinguishable. Any
Float will answer true to .isInstanceOf[Double], and vice versa.

This is fine, because floats and doubles already behave exactly
the same way when they are primitives, since they are both
represented as numbers and all operations on them are the direct
equivalent on number.

Double is slightly privileged compared to Float in the sense that
x.getClass() will return classOf[java.lang.Double] (and not Float)
when x is a non-integer number.