New intermediate representation (IR).
This new IR is meant to be an interchange format between the
compiler and the packager and optimizers (instead of raw
JavaScript code), so that inlining will eventually be possible.
In this commit, though, it simply replaces the JS Trees as an
intermediate representation in the compiler: the compiler
emits the IR and immediately desugars it in the JavaScript
subset, which is emitted as source code as before.

Some highlights of the IR compared to JS trees:
* It represents Scala.js classes independent of an encoding
  (so the variable `ScalaJS` and its fields are never in the IR)
* It is statically typed, with an additional `dyn` type.
  (note: I said it is *statically typed*; I did not say it was
  *typechecked*!)