Change the API and serialization format of class info.
In 0.4.x, class info were stored in .sjsinfo files with a JSON
format. In the initial transformations to serialize the IR, we
have limited the changes and still produced class info as an
IR tree with the exact same JSON structure as before.

However, this format of serialization is unnecessary complex and
slow to parse, which is a pity since class info (and even more
importantly the subset of "rough class info") should be parsed
quickly.

This commit introduces a binary format for serializing the class
info, and in the process lifts the class info as a proper API in
the IR package.