Give true (hard-coded) IR to java.lang.{Object,String}.
These were the 2 only classes in the universe that had to be
special-cased in various ways because they did not have true
IR, most notably in the optimizer. It would probably have gotten
worse in the inliner.

With this commit, the only hand-written JavaScript that we still
have is the corejsenv.js, for which it is reasonable, as it does
not take part in Scala.js-specific optimizations.