Add incremental inliner.
The incremental inliner is run as part of fastOptJS. Because it
is incremental, it runs very quickly starting from the second
run on the same codebase when changes are made.

The inliner *reduces the size* of the emitted code by a small
margin, and speeds up the program execution by 1.3x to 3.3x,
according to the benchmarks we ran.