Fix #825: ParamDefs of tailrec methods should be mutable.
Considering them as immutable was fine for the purposes of
JSDesugaring, but not for the inliner, causing subtle bugs.

This commit simulates the binary compatibility scenario by
not changing the compiler, and relying on the deserializer
to apply hacks to fix old IR. A follow up commit will
update the compiler to emit correct IR directly.