Disable optimistic optimizations if there are too many rollbacks.
This is a protection mechanism against the potential exponential
blow-up of tryOrRollback, in the general case. If, during the
optimization of a given method, more than 256 rollbacks occur,
then we bail out completely, cancel everything, and start the
optimization of the method from scratch, but this time disabling
any optimistic optimizations.