Fix binary version for snapshots of minor releases.
The binary version for 0.5.3-SNAPSHOT should in fact be 0.5,
not 0.5.3-SNAPSHOT. However, the binary version for
0.6.0-SNAPSHOT, as well as 0.6.0-M1 and so on, should be the
full version indeed. In general, snapshots of minor releases
should have the binary version of their major release. This
is important for applications built on a snapshot to be able
to resolve libraries built with a release.

In the process, this commit also cleans up a few things
related to version numbers in the sbt plugin.