Use final vals where possible in javalanglib.
This means in all the companions of boxed classes. This
makes the constructors of these companions elideable.

The values Float.MIN_NORMAL and Double.MIN_NORMAL have been
deleted, because there was no way to make them constant
expressions. Since they are final vals anyway in the JDK, they
will be constant-folded by scalac, so there is no need to emit
them in our library.