Aggressively inline methods when arguments are "likely" optimizable.
If at least one of the receiver and arguments to a method call
is "likely optimizable", then we force inlining of the method,
even if it is not inlineable.

An argument is likely optimizable if it is a LocalDef for a
tentative anon fun replacement, or for an inlined class instance
(i.e., a record var ref).

At the same time, we removed ad-hoc @inline's on methods, because
they were targeted at the use case now better covered by this
improvement.

Together, these changes perform less inlinings that serve no
purpose, and more inlinings that do bring improvements.