Reduce strain on CI
This moves 8 tasks from the PR CI configuration to the nightly and
moves most partest tasks from the nightly to a new weekly build.

N.B.: We should not be doing this, but given the load on the CI we
probably have no choice. The full tests should be introduced ASAP.